fix
Fix order of operations in KMProfileRoutes constructor

The constructor of the `KMProfileRoutes` class has been modified to assign the `parent` parameter to the private field `p` before calling the `CreateRoutes()` method. This change ensures that the `p` field is initialized before any routes are created, which might be necessary if `CreateRoutes()` relies on `p` being properly initialized.